Working version of ASDM v2
