+++ BuiltInTestModels/Stella_outcomes/Basic_math.csv
@@ -1,0 +1,10 @@
+Months,a,b,c
+0,2,3,5
+0.25,2,3,5
+0.5,2,3,5
+0.75,2,3,5
+1,2,3,5
+1.25,2,3,5
+1.5,2,3,5
+1.75,2,3,5
+2,2,3,5
