+++ BuiltInTestModels/Stella_outcomes/Non-negative_stocks_with_flows.csv
@@ -1,0 +1,9 @@
+Months,Converter 1,Converter 2,Converter 3,Flow 1,Flow 2,Flow 3,Flow 4,Flow 5,Flow 6,Flow 7,Flow 8,InFlow,InFlow2,OutFlow 1,OutFlow 2,Stock 1,Stock 2,Stock 3,Stock 4,Stock 5
+1,50,10,40,15,5,15,5,20,10,20,10,30,0,50,10,40,0,20,20,20
+2,40,0,10,0,0,0,0,20,10,10,0,30,0,40,0,10,0,0,0,0
+3,30,0,0,0,0,0,0,20,10,10,0,30,0,30,0,0,0,0,0,0
+4,30,0,0,0,0,0,0,20,10,10,0,30,0,30,0,0,0,0,0,0
+5,30,0,0,0,0,0,0,20,10,10,0,30,0,30,0,0,0,0,0,0
+6,30,0,0,0,0,0,0,20,10,10,0,30,0,30,0,0,0,0,0,0
+7,30,0,0,0,0,0,0,20,10,10,0,30,0,30,0,0,0,0,0,0
+8,30,0,0,0,0,0,0,20,10,10,0,30,0,30,0,0,0,0,0,0
