+++ BuiltInTestModels/Stella_outcomes/Time_unit.csv
@@ -1,0 +1,10 @@
+Year,a
+1,1
+1.25,1.25
+1.5,1.5
+1.75,1.75
+2,2
+2.25,2.25
+2.5,2.5
+2.75,2.75
+3,3
