+++ BuiltInTestModels/Time_unit.stmx
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <xmile version="1.0" xmlns="http://docs.oasis-open.org/xmile/ns/XMILE/v1.0" xmlns:isee="http://iseesystems.com/XMILE">
 	<header>
-		<smile version="1.0" namespace="std, isee" uses_conveyor=""/>
-		<name>Conveyor</name>
-		<uuid>d3f11430-2483-4912-9db4-e462f7beb286</uuid>
+		<smile version="1.0" namespace="std, isee"/>
+		<name>Time_unit</name>
+		<uuid>03d2a1c7-c496-46b8-a80b-1d72b8a23b23</uuid>
 		<vendor>isee systems, inc.</vendor>
 		<product version="2.0" isee:build_number="2176" isee:saved_by_v1="true" lang="en">Stella Professional</product>
 	</header>
-	<sim_specs isee:sim_duration="0.5" isee:simulation_delay="0.1" isee:restore_on_start="false" method="Euler" time_units="Months" isee:instantaneous_flows="false" isee:loop_scores="false" isee:loop_exhaustive_allowed="1000">
-		<start>0</start>
-		<stop>5</stop>
+	<sim_specs isee:sim_duration="0.5" isee:simulation_delay="0.25" isee:restore_on_start="false" method="Euler" time_units="Year" isee:instantaneous_flows="false" isee:loop_scores="true" isee:loop_exhaustive_allowed="1000">
+		<start>1</start>
+		<stop>3</stop>
 		<dt reciprocal="true">4</dt>
 	</sim_specs>
-	<isee:prefs show_module_prefix="true" live_update_on_drag="true" show_restore_buttons="false" layer="model" interface_scale_ui="true" interface_max_page_width="10000" interface_max_page_height="10000" interface_min_page_width="0" interface_min_page_height="0" rerun_on_structure_change="false" saved_runs="1" keep="false" rifp="true"/>
+	<isee:prefs show_module_prefix="true" live_update_on_drag="true" show_restore_buttons="false" layer="model" interface_scale_ui="true" interface_max_page_width="10000" interface_max_page_height="10000" interface_min_page_width="0" interface_min_page_height="0" rerun_on_structure_change="false" saved_runs="5" keep="false" rifp="true"/>
 	<isee:multiplayer_settings include_chat="true" allow_observers="false" advance_time_increment="1" observer_start_page="home_page" enabled="false"/>
 	<isee:time_formats default_format="Builtin">
 		<isee:time_format name="Builtin" type="adaptive"/>
@@ -23,38 +23,11 @@
 			<all/>
 		</export>
 	</data>
-	<model_units>
-		<unit name="Months">
-			<eqn/>
-			<alias>mo</alias>
-			<alias>month</alias>
-		</unit>
-		<unit name="Per Month">
-			<eqn>1/Months</eqn>
-		</unit>
-	</model_units>
+	<model_units/>
 	<model>
 		<variables>
-			<stock name="Conveyor 1">
-				<eqn>12</eqn>
-				<inflow>Flow_In</inflow>
-				<outflow>Flow_Out</outflow>
-				<conveyor>
-					<len>TransitTime</len>
-				</conveyor>
-			</stock>
-			<flow name="Flow Out">
-				<eqn>0</eqn>
-				<non_negative/>
-				<units>Per Month</units>
-			</flow>
-			<flow name="Flow In">
-				<eqn>1</eqn>
-				<non_negative/>
-				<units>Per Month</units>
-			</flow>
-			<aux name="TransitTime">
-				<eqn>4</eqn>
+			<aux name="a">
+				<eqn>TIME</eqn>
 			</aux>
 		</variables>
 		<views>
@@ -139,36 +112,7 @@
 					</graph>
 					<table color="black" background="#E0E0E0" text_align="right" font_size="12pt" orientation="vertical" wrap_text="false" isee:auto_fit="true" isee:use_alternate_row_colors="false" isee:unlimited_table_length="false" blank_column_width="80" column_width="160" interval="1" report_balances="beginning" report_flows="instantaneous" header_font_style="normal" header_font_weight="normal" header_text_decoration="none" header_text_align="center" header_vertical_text_align="center" header_font_color="black" header_font_family="Arial" header_font_size="12pt" header_text_padding="2" header_text_border_color="black" header_text_border_width="thin" header_text_border_style="none"/>
 				</style>
-				<stock x="470" y="116" name="Conveyor 1"/>
-				<flow x="582.75" y="118" name="Flow Out">
-					<pts>
-						<pt x="492.5" y="118"/>
-						<pt x="697" y="118"/>
-					</pts>
-				</flow>
-				<flow x="352.75" y="118" name="Flow In">
-					<pts>
-						<pt x="258" y="118"/>
-						<pt x="447.5" y="118"/>
-					</pts>
-				</flow>
-				<stacked_container uid="1" x="228" y="181" width="514" height="344">
-					<graph width="514" height="344" type="time_series" isee:points_only="false" show_grid="false" include_units_in_legend="false" plot_numbers="false" isee:label_pie_slices="false" num_x_grid_lines="0" num_y_grid_lines="0" num_x_labels="6" num_y_labels="21" isee:fill_intensity="0.1" isee:allow_zero_axis="true" left_axis_multi_scale="false" left_axis_auto_scale="true" left_include_units="true" right_axis_multi_scale="false" right_axis_auto_scale="true" right_include_units="true">
-						<plot color="blue" isee:keep_zero_visible="false" pen_width="1" index="0" show_y_axis="true">
-							<entity name="Flow_In"/>
-						</plot>
-						<plot color="red" pen_style="dot_dashed" isee:keep_zero_visible="true" pen_width="1" index="1" show_y_axis="true">
-							<entity name="Conveyor_1"/>
-						</plot>
-						<plot color="fuchsia" pen_style="dotted" isee:keep_zero_visible="true" pen_width="1" index="2" show_y_axis="true">
-							<entity name="Flow_Out"/>
-						</plot>
-						<plot color="#008F44" pen_style="dashed" isee:keep_zero_visible="true" pen_width="1" index="3" show_y_axis="true">
-							<entity name="TransitTime"/>
-						</plot>
-					</graph>
-				</stacked_container>
-				<aux x="367" y="30" name="TransitTime"/>
+				<aux x="183" y="174" name="a"/>
 			</view>
 		</views>
 	</model>
